pyproject.toml
[project]
name = "trace"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "google-genai>=1.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "cryptography>=46.0.7",  # CVE-2026-39892 fix; transitive via google-genai
    "markitdown[pdf]>=0.1.0",  # PDF/URL → clean Markdown for crawl_single.py
    "stix2-validator>=3.2",  # OASIS cti-stix-validator
    "pyyaml>=6.0",  # input/sources.yaml parsing
    "python-dotenv>=1.0",  # load .env into os.environ at startup
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=9.0.3",
    "pytest-cov>=5.0.0",
    "pip-audit>=2.7.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not integration'"
markers = [
    "integration: marks tests that require real Vertex AI (run with: pytest -m integration)",
]
